fix: namespace mapping fix
Namespace mapping was not occurring in a vital place, this has been rectified to stop us creating multiples of the same identifier (eissn AND issn for example) and then matching on them as if they were the same, giving us a double match and a failure.

ERM-1847 - This will stop the data problem which causes that crash. Separate work done to ensure no crash in this case in future