Release 5.4.2 (#614)
* fix: ExternalEntitlements (#606)

External entitlements no longer fetch resource on `index` action. In addition, subscriptionAgreement now accepts queryParam `excludes` to allow selective exclude from the query itself

ERM-2423

* fix: fetchExternalResources (#607)

Added a queryParam to quash the fetching of external resources. It's a little clunky because it needed to keep backwards compatibility, but there may be a rewrite of this stuff in Orchid and beyond, so for Nolana bugfix it's the most expedient thing to do

ERM-2423

* Fix/entitlement json errors (#611)

* fix: Remove duplication in JSON

Ensure the render method excludes any properties that are handled
explicitly by the supplied Closure.

* fix: Template inheritance causes weird empty object

Extend the template specifically as the g.render seems to be causing
issues.

* fix: Swapped out alias for nested exists subqueries (#612)

Swapped out the createAlias solution for only displaying TIs with attached PCIs down the line for a nested exists subqueries solution.  Seems to be equally/more performant and solves the issue

ERM-2443

* refactor: ResourceController (#613)

* refactor: ResourceController

Use Constants interface to 'nicely' import Query constant 'this' to help mitigate future scenario wherre this needs changing in WeTK

* refactor: Change import

Directky import Constants.Queries

* chore: Release 5.4.2

Co-authored-by: sosguthorpe <steve.osguthorpe@k-int.com>