fix: domain class fetching
Grails 5 has tightened security on fetching from within a domain class on the view. In reality we should stop this whenever possible.

However, in some places this will be utterly non-performant, and in others we simply don't have a "proper" domain class link so need a fetch. Those cases we do the work within a Tenants.withCurrent to ensure hibernate sessions are observed.

In addition, BaseSpec has changed so that access to specDefaultHeaders -> state.specDefaultHeaders