fix: ERM-3177: Check on `noSICount` in WorkSourceIdentifierTIRSImpl is not correct
Previously WorkSourceIdentifierTIRS used check `work.sourceIdentifier = null` in the HQL statement, assuming that the Hibernate query language and GORM would play nicely for null check. This was incorrect however, so it has been swapped for a more direct count, of the number of Works LEFT JOIN the IdentifierOccurrences, where the IdentifierOccurrence is null.

NOTE: This will only work so long as Work only has a single IdentifierOccurrence attached to it in this manner. Reworks will be needed if this changes.

ERM-3177