fix: Changed transaction handling in JobRunnerService (#754)
Spring @Transactional annotation does not apply on `protected` or `private` methods. Previously Grails 5 either permitted this or this being ignored did not cause us an issue. However with grails 6 the handling of transactions has changed, and so we need to explicitly wrap the execution of the job in a GormUtils withNewReadOnlyTransaction.

Also increased the heap size in build.gradle for developers running the module using bootRun, to 1.5G

ERM-3190