fix: Only allow electronic returns from classOneMatch
Previously classOneMatch assumed it would only ever find electronic class one matches.

This is not always the case, if an incoming electronic citation has an issn which in the database exists only on a print title.

In general this is likely a data issue, however in our case this caused the TIRS logic to flow forwards assuming we were working with an electronic title, and not a print sibling of an electronic title.

In import this caused bad data output, and in ingest this locked up a database session (differences due to slight differences in transaction/session set ups for the two ingress methods)

The fix ensures that classOneMatch _only_ returns electronic titles, and thus in the case above the logic flows onto "siblingMatch" and works as expected

ERM-3166