perf: Tweak listDeduplicator
This is unlikely to be _much_ of a performance gain, but listDeduplicator previously ran in O(n^2) time, and this should run in O(n) time but lose ordering (which we shouldn't care about). However 90% of the time this list was only 1 or 2 ids long, so I'm not expecting a very large speed increase from this.