fix: WIP--buildIdentifierHQL
This is a change to how buildIdentifierHQL will work -- for now it is an _additional_ method, which the TIRSs have moved over to, but the original code should be removed before merging.

The previous buildIdentifierHQL was flawed, in that it would have automatically returned _null_ for a direct match on any more than a single identifier. This was accidentally fine since the HQL was only ever used for sibling matching, but would have caused us issues if that ever changed.

This HQL should be functionally equivalent for all our use cases, but may carry a performance change (difficult to measure that though, could honestly be either slower or faster)