ERM-3281: Design and implement process to resolve eissn, pissn, eisbn and pisbn namespaces (#792)
* build: Bump the dev machine maxpoolsize to 50

Bumped default maxpoolsize for rancher desktop and docker compose profiles to 50 from 10

* perf: Tweak listDeduplicator

This is unlikely to be _much_ of a performance gain, but listDeduplicator previously ran in O(n^2) time, and this should run in O(n) time but lose ordering (which we shouldn't care about). However 90% of the time this list was only 1 or 2 ids long, so I'm not expecting a very large speed increase from this.

* fix: WIP--buildIdentifierHQL

This is a change to how buildIdentifierHQL will work -- for now it is an _additional_ method, which the TIRSs have moved over to, but the original code should be removed before merging.

The previous buildIdentifierHQL was flawed, in that it would have automatically returned _null_ for a direct match on any more than a single identifier. This was accidentally fine since the HQL was only ever used for sibling matching, but would have caused us issues if that ever changed.

This HQL should be functionally equivalent for all our use cases, but may carry a performance change (difficult to measure that though, could honestly be either slower or faster)

* refactor: DirectMatch

Moved directMatch to BaseTIRS so it's available everywhere as a public method. Include ability to swap between matching _any_ of the passed identifiers (default) or _all_ of the passed identifiers with a mustMatchAll prop

TODO -- should probably test this directly in BaseTIRSSpec

* chore: Comment

* refactor: IdentifierException

Added IdentifierException to allow us to do specific logic on IdentifierExceptions between IdentifierService and TIRSs

Tweaked comments in some TIRSs (not really relevant to anything)

* refactor: Moved identifier lookup logic to service

Moved lookup and create of identifiers/namespaces to IdentifierService, with BaseTIRS able to catch and rethrow any understood exceptions as TIRSExceptions instead

* chore: Comment changes

* chore: Exception handling, should only rethrow the IdentifierExceptions we can understand as TIRSExceptions, leaving the rest as is

* fix: There was a bug in the lookupOrCreateIdentifier exception throwing...

Previously tried to throw an exception with a message containing a non-existing object...whoops. (I don't think it ever got called but hoo boy)

* refactor: IdentifierService method defaults

- Made lookupOrCreateIdentifier a public method
- Swapped to default "flush: false" instead, with the TIRS insisting on a flush to maintain logic workflow as before

* test: IdentifierServiceSpec

Added withTenant closure call to BaseSpec to avoid Tenants logic everywhere
First blush at an integration test for some of the functionality of IdentifierService
Also added an exception code to IdentifierException

* chore: Import whoops

* feat: fixEquivalentIds

First attempt at fixEquivalentIds, along with "happy path" test case in IdentifierServiceSpec (more totest there though)

* refactor: "Prime occurrences" wrangling

Handle cases where we go to move an identifierOccurrence over to the "prime identifier", but there is already a "prime occurrence" for that resource. In these cases be careful with statuses and then remove the current identifier entirely

* test: Added test case for IdentifierException throwing

* test: Check fixEquivalentIds occurrence wrangling

- Ensure that occurrences are checked, and deleted where there was previously an occurrence in place, but changed where there was not. In addition make sure that statuses behave as expected

- Test found a bug (Who'd have thought it) -- fixed that. Switch case falling through where it shouldn't

- Test found a bug 2, electric boogaloo -- fetching refdata value with dynamic lookup in a loop caused entity access exception. Using HQL here _would_ potentially work, allowing hibernate to manage its proxies, but that gets messy if we need to create error. To avoid this entirely just fetch "approved" status once at top of method and apply at will within the loop.

* perf: Small identifier wrangling performance boost

Only check approved TI IdentifierOccurrences when looking for ones to set to ERROR in updateIdentifiers in WorkSourceIdentifierTIRS

Also small comment for clarity in IdFirstTIRS and removed TODO from IdentifierService

* fix: First attempt at using fixEquivalentIds in IdFirstTIRS

First attempt at integration of fixEquivalentIds in IdFirstTIRS (necessary in fallback from WorkSourceIdTIRS)
- Removed limit on fetched Identifiers (should only ever see 2 at most anyway even with bad data, but this allows us to catch any rare cases with 3+)
- Slightly tweak logic when matched id is found, no longer allow
- Tweaked fixEquivalentIds logging to output the ids being equated

* test: First int test work for fixEquivalentIds

Added initial test setup for fixEquivalentIds to IdFirstTIRSSpec

* chore: Comment and remove commented code

* chore: Identifier toString

Added default toString to Identifier of "${ns.value}:${value}". Made tweaks in a couple of logging places to make use of this

* test: Finish test case for IdFirstTIRS making use of fixEquivalentIdentifiers (happy path)

* test: Error test cases

Test cases for exceptions thrown in "fixEquivalentIds" part of IdFirstTIRS