perf: Coverage calculations
ErmResource previously _always_ calculated coverage on save (in beforeValidate). This is normally desired behaviour (if not the best way to achieve this) because we want any save anywhere in the heirachy to propogate. However during ingest coverage is expected to be set manually, and this causes extra looping behaviour for no benefit. FOR NOW implement workaround where a transient can prevent this additional calculation step on save