ERM-3308: Calculate coverage performance tweaks (#799)
* perf: Coverage calculations

ErmResource previously _always_ calculated coverage on save (in beforeValidate). This is normally desired behaviour (if not the best way to achieve this) because we want any save anywhere in the heirachy to propogate. However during ingest coverage is expected to be set manually, and this causes extra looping behaviour for no benefit. FOR NOW implement workaround where a transient can prevent this additional calculation step on save

* fix: Coverage save propogation

Ensure last save makes it through regular propogation checks

* perf: Coverage queries

Rewrite a couple of internal coverage queries as HQL to try and imrpive performance (Seems to have a small benefit)

* fix: calculateCoverage tweak

Small tweak to ensure calculateCoverage is _always_ called at the end of the ingest process, don't hide that in the manual coverage setting but turn it off there and ensure it's called at the end

* ci: Tweak to run-int-tests

Swap out docker-compose for docker compose

* fix: utilityService.checkValidBinding oversight

Was previously calculating coverage for every coverage statement coming in via the utilityService.checkValidBinding. This obviously triggers beforeValidate, so turn off calculateCoverage there too