fix: Migrations pre-breaking GIN index insertion (#818)
Inserted migration files prior to `update-mod-agreements-6-0.groovy` and `add-quesnalia-indices-and-fk-constraints.groovy` files. These contain index definitions set up _before_ the non-trigramised (not sure if that's a word) versions in those files. This _should_ allow institutions who have already run these migrations to continue to work with the indices in place (Although we heavily recommend an operational task to bring those in line with the "tip of master" code, while allowing those institutions who are currently prevented from upgrading to do so as normal.

This is a slight misuse of migrations to accommodate this unfortunate gap between institutions who have already upgraded and those who have not.

NOTE: This does _not_ fix the issue where column `sa_name` on `subscription_agreement` table continues to not have a GIN index thanks to this column already having an index on that column from uniqueness constraint.

ERM-3321