feat: Stack traces should not be included in API responses. Refs ERM-3292 (#890)
* feat: Set up module to use centralised error handling from webtoolkit (WIP, WTK not released yet)

This involves removing the 500 routing in UrlMappings.groovy, as well as the `error.gson` view file.

These can be reinstated for overriding either _just_ the view, and leaving the error logic in the plugin controller, or the full routing.

In addition configure the application.yml for prod with `endpoints.include-stack-traces: false`, and application-dc.yml with `endpoints.include-stack-traces: true` so that dev builds continue to include stacktraces in responses

refs [ERM-3292](https://folio-org.atlassian.net/browse/ERM-3292)

* docs: Include ENDPOINTS_INCLUDE_STACK_TRACE in README file

* build: Updated web-toolkit-ce to 10.2.0

* chore: Actually remove error urlMapping

ERM-3292: Stack traces should not be included in API responses (#890)