fix: MultipleAliasSQLCriterion didn't play nice when a query was re-called (such as during totalCount)
Previously aliases were fetched using the getSQLAlias method directly, which assumes the hash of the query will always remain the same.

This was not the case during totalCount calls, causing an alias to go missing and crash the query

criteriaQuery does not directly allow access to its alias maps, but the issue can be forced by appending a "." to the end of alias, thus forcing it to lookup via the known alias instead of the criteria.

This API is horrible and flaky, but unlikely to change as it's deprecated anyway. I still hate it