feat: Ownership cycle (#911)
* feat(WIP): MultipleAliasSQLCriterion addition

Added ability to house multiple aliases within a single HibernateCriteriaBuilder criteria for use in SqlRestriction. HOWEVER, this does not work with doTheLookup stats, the COUNT query breaks :(

I don't think this has a future to be honest, would need to check with Ian/Steve

* fix: MultipleAliasSQLCriterion didn't play nice when a query was re-called (such as during totalCount)

Previously aliases were fetched using the getSQLAlias method directly, which assumes the hash of the query will always remain the same.

This was not the case during totalCount calls, causing an alias to go missing and crash the query

criteriaQuery does not directly allow access to its alias maps, but the issue can be forced by appending a "." to the end of alias, thus forcing it to lookup via the known alias instead of the criteria.

This API is horrible and flaky, but unlikely to change as it's deprecated anyway. I still hate it

* fix: Multi-level ownership

Second tier of ownership was not obeying proper ordering. Instead make sure we use the top level alias

* refactor: Centralise PolicyControlled

Annotations are not extensible in the way I thought they were, so move to using the central annotation instead, and then we can handle core logic a little cleaner in the central library

* refactor: Clean up ownership trawling so majority of logic lives in PolicyControlledManager

Column/Field are now BOTH tracked by the PolicyControlled

* docs: Added javadoc comments

Also tweaked import path for special criteria work

* fix: Slight break in policySqlFragments code

Also more javadoc work