feat: Hibernate parameters
Change over SQL from raw String to AccessControlSql. This allows us to make use of Hibernate's parameter work to avoid SQL injection.

In order to accomplish this, we need both to spin up AccessControlSqlType enum to internally track what types our parameters have, and then in the grails mapping to create AccessControlHibernateTypeMapper.

This keeps the hibernate logic in the grails layer, away from AccessControl, and allows us to map to Hibernate types in MultipleAliasSQLCriterion, AND in nativeQuery.