fix: Change AccessPolicy in core to interface instead of class
Ensure that each implementation above (Grails/MN etc) is forced to spin up its own Entity class which actually owns all of the properties on it. This is important in Grails in particular because the "dirtyCheck" magic was not being applied to the properties, and so edits would not take.

Also ensure that special case delete in SubscriptionAgreementController is access controlled