fix: Re-remove the extra permissive clause
At the minute there seems to be a discrepancy between behaviour on READ and on EDIT for acquisition units.

On READ it is currently enough for a non-restrictive unit to be present on that resource. For EDIT however the same scenario leads to a restriction of the functionality

For now, leave the scaffolding in place to support the former, but assume that the latter behaviour is the expected one