refactor: Set up RestrictionTree skeleton which lives on the PolicyControlledManager
This pattern allows us to store the mapping per restriction instead of per-level. We can then "enrich" this in the policyEngine with parameters and subqueries at each level.

This provides us with an object which we can build any way wee like to get SQL back for a given restriction