feat: Resource access policies do not protect agreement line creation -- ERM-3862 (#957)
* feat: PolicyControlled Restriction owner mappings

Added a mapping from each level of the ownership chain in PolicyControlledManager, corresponding to how that level handles a given restriction

(options include pass through to owner, use standalone policies, both, or neither)

Added test assertion to that effect in PolicyControlledManagerTest

* fix: Undo accidental TODO change, not necessary

* fix: Javadoc failed

* feat(PolicyControlledManager): Add helper functions for getting a particular owner level safely and returning the size of the ownership chain

* feat(RestrictionMapEntry): Enhance builder with fluent accessors and maintain equality/hashcode

* feat(PolicyControlledValidator): Set up Validator class to validate uses of @PolicyControlled

Newly mandatory: resourceIdColumn, resourceIdField and resourceTableName

Test updates to match

* feat(PolicyControlledManager): Ability for PolicyControlledManager to spit out SQL for obtaining a given owner level's id

* docs(PolicyControlledManager): javadoc extension

* feat(AccessPolicyAwareController): !WIP! Set up ownership chain parsing for getPolicySql

* fix(session): Fix session management to marry up to changes on master

* refactor: Start ownership level at 0 instead of -1 for the policyControlledManager ownership chain

I _think_ this is no longer needed thanks to the SQL creation built into the policyControlledManager

* fix: Previous changes had broken the CLAIM operation

This does NOT introduce the changes required to have standalone CLAIM policies on every level, only to fix what had been previously broken

* chore: Include a clause label in the SQL generated by AcquisitionUnitPolicySubquery

ALSO... note that this is creating SQL via DIRECT string SQL injection. It must be INCREDIBLY careful to only set these parameters via trusted code.

Unfortunately this is somewhat unavoidable when trying to set up dynamic SQL. Attempt to mitigate by avoiding any dodgy SQL characters (hopefully)

* fix: More SQL injection protection -- now allows for hibernate style {alias} as well

* fix: More SQL injection protection -- now allows for hibernate style {alias} as well

* feat: Swap the building of SQL up the hierarchy over to an object oriented methodology

* refactor: Move the bulk of the RestrictionTree creation logic and subquery fetching logic at each level over to core components

* refactor: Set up RestrictionTree skeleton which lives on the PolicyControlledManager

This pattern allows us to store the mapping per restriction instead of per-level. We can then "enrich" this in the policyEngine with parameters and subqueries at each level.

This provides us with an object which we can build any way wee like to get SQL back for a given restriction

* feat: Add ability to fetch subqueries for multiple restrictions at once

* feat(AcquisitionsClient): Added ability to fetch all protected pairs in single fetch

* refactor: Swap over to fetching restrictive AND non-restrictive units for a restriction in a single fetch

Still todo -- multi-restriction fetch amalgamation

* refactor: AcquisitionsClient can now handle fetching user acquisition units for multiple different restrictions at once only making 2 calls instead of 3*number_of_restrictions

Performance seems kinda bad still, but I think that Sunflower is chugging... remains to be seen on snapshot

* refactor: Tweak the provider pattern to instead pass an ERTParameterProvider implementation

Still not happy with this at the moment, but this will allow me to move onto CREATE

* refactor: GrailsERTParameterProvider

Added Grails implementation of the ERTParameterProvider interface which we can build with a policyControlledManager, resourceId and startLevel, so that it can be built per request and we don't have to pass all of that information into the PolicyEngine

* feat: CREATE is now mapped from the owner in the request body

* refactor: Some code moving and documentation writing

* feat: CLAIM now works at any level in the ownership tree

* feat: Policies fetch now returns mapped list of all policies for all ownership levels

Needs refactors

* refactor: Minor refactors... more needed

* feat: Delete now also handles any AccessPolicyEntity objects for the deleted resource

Also refactored and renamed a bunch of stuff

* docs: Javadocs

* docs: Javadoc comments

* test: Add test for AcquisitionUnitPolicyEngine

* feat: Make use of new PolicyControlled features for entitlement control

* fix: Re-allow orphaned owner CREATE of Entitlements. This may need revisiting though -- see comments

* fix(AcquisitionUnitPolicySubquery): SQL should return 1=1 for TRUE not just 1

* fix: Allow for empty subqueries